angular-advanced/solutions/unit-test/exercise-6/src/app/login/login.component.ts
import { Component } from '@angular/core';
import { FormBuilder, Validators } from '@angular/forms';
import { Router } from '@angular/router';
import { AuthService } from '../service/auth.service';

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss']
})
export class LoginComponent {

  fg = this.fb.group({
    email: ['', [Validators.required, Validators.email]],
    password: ['', Validators.required],
  });

  constructor(private fb: FormBuilder, private authService: AuthService, private router: Router) { }

  login(): void {
    const { email, password } = this.fg.value;
    if(this.fg.valid) {
      this.authService.login(email!, password!).subscribe(value => {
        if(value) {
          this.router.navigate(['/'])
        }
      });
    }
  }

}